requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
numpy<2.0
faiss-cpu==1.8.0.post1
pypdf==5.0.1
openai==1.50.2
httpx==0.27.2
